Improve comment accuracy
Cache line bouncing usually refers to multiple cores writing to
(different parts of) the same cache line. Each core will need to
gain exclusive access to the cache line. Thus, the ownership of
of the cache line will be bouncing among the cores that want to
modify it.

Here, the reader loads `produced_[head_]` and the writer modifies
`produced_[tail_]`. If `tail_` and `head_` are close enough (in
the same cache line), the cache coherence protocol will force
the cache from the reader's core to be invalidated upon a write.
And this cache invalidation takes extra time, which also makes
the writer slower.